+++ spec/spec_helper.rb
@@ -40,7 +40,7 @@ end
 
 module DummyCreator
   def self.id
-    "4321"
+    "4a3f2e1d"
   end
 end